Refactor remote runners into a base + "class per runner".
This reduces coupling (using fabric is just an implementation detail of the
remote runner) and nasty ifs and checks inside the FabricRunner class we had
before.

This will also makes it easier to move away from Fabric runner in the future.
