+++ e1.yaml
@@ -1,0 +1,18 @@
+    name: "examples.webhook_file1"
+    description: "Sample rule dumping webhook payload to a file."
+    enabled: true
+
+    trigger:
+        type: "core.st2.webhook"
+        parameters:
+            url: "sample"
+
+    # criteria:
+    #     trigger.body.name:
+    #         pattern: "st2"
+    #         type: "equals"
+
+    action:
+        ref: "core.local"
+        parameters:
+            cmd: "echo \"{{trigger.body}}\" >> /tmp/st2.webhook_sample.out"
