+++ st2tests/st2tests/fixtures/executions/runners.yaml
@@ -9,7 +9,7 @@ run-local:
   description: A runner to execute local actions as a fixed user.
   enabled: true
   name: run-local
-  runner_module: st2actions.runners.fabricrunner
+  runner_module: st2actions.runners.localrunner
   runner_parameters:
     cmd:
       type: string
